# qzt-go-server/cmd/server/docs/swagger.yaml
basePath: /
definitions:
...
  service.ApproversDesign:
    properties:
      approval_type:
        type: string
      approver_direction:
        type: string
      approver_list:
        type: string
      approver_type:
        type: string
      cc_list:
        type: string
      cc_type:
        type: string
      empty_approver_action:
        type: string
      fallback_approver:
        type: integer
      multi_approver_mode:
        type: string
      node_number:
        type: string
      same_submitter_action:
        type: string
    type: object
...
  service.ConditionsDesign:
    properties:
      condition_config:
        type: string
      node_number:
        type: string
    type: object
...
  service.CreateFlowRequest:
    properties:
      enable:
        type: integer
      form_key:
        type: string
      form_type:
        type: string
      name:
        type: string
      number:
        type: string
    required:
    - form_type
    - name
    type: object
...
  service.CreateSysJobRequest:
    properties:
      bean_class:
        type: string
      cron_expression:
        type: string
      job_group:
        type: string
      job_name:
        type: string
      remark:
        type: string
      status:
        type: integer
    required:
    - bean_class
    - cron_expression
    - job_name
    type: object
...
  service.LinkDesign:
    properties:
      from_node_number:
        type: string
      sort:
        type: integer
      to_node_number:
        type: string
    type: object
...
  service.NodeDesign:
    properties:
      execute_timing:
        type: string
      name:
        type: string
      node_type:
        type: string
      number:
        type: string
      sort:
        type: integer
    type: object
...
  service.SaveDesignRequest:
    properties:
      approvers:
        items:
          $ref: '#/definitions/service.ApproversDesign'
        type: array
      conditions:
        items:
          $ref: '#/definitions/service.ConditionsDesign'
        type: array
      links:
        items:
          $ref: '#/definitions/service.LinkDesign'
        type: array
      nodes:
        items:
          $ref: '#/definitions/service.NodeDesign'
        type: array
    required:
    - nodes
    type: object
...
  service.SyncFeaturesRequest:
    properties:
      item_ids:
        items:
          type: integer
        type: array
      module:
        type: string
    required:
    - module
    type: object
  service.ToggleModuleRequest:
    properties:
      enabled:
        type: boolean
      module:
        type: string
    required:
    - module
    type: object
...
  service.UpdateSysJobRequest:
    properties:
      bean_class:
        type: string
      cron_expression:
        type: string
      job_group:
        type: string
      job_name:
        type: string
      remark:
        type: string
      status:
        type: integer
    required:
    - bean_class
    - cron_expression
    - job_name
    type: object
...
  xresponse.Response:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      msg:
        example: success
        type: string
      timestamp:
        type: integer
    type: object
host: localhost:9000
info:
  contact: {}
  description: |-
    基于 Go 的模块化单体后端(基础平台 + RBAC + 系统管理)。第一阶段地基。
    统一响应信封 {code,msg,data,timestamp};登录返回 JWT access/refresh 双令牌。
  title: 企业级业务管理平台 API
  version: "1.0"
paths:
...
  /approval/flows:
    get:
      parameters:
      - description: 页码(默认1)
        in: query
        name: page
        type: integer
      - description: 每页条数(默认10)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/xresponse.Response'
      security:
      - BearerAuth: []
      summary: 审批流程列表
      tags:
      - 审批流程
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建流程请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.CreateFlowRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/xresponse.Response'
      security:
      - BearerAuth: []
      summary: 创建审批流程
      tags:
      - 审批流程
  /approval/flows/{id}:
    get:
      parameters:
      - description: 流程ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/xresponse.Response'
      security:
      - BearerAuth: []
      summary: 审批流程详情
      tags:
      - 审批流程
  /approval/flows/{id}/design:
    put:
      consumes:
      - application/json
      description: 保存审批节点图(节点/审批人/条件/连线),每次保存创建新版本
      parameters:
      - description: 流程ID
        in: path
        name: id
        required: true
        type: integer
      - description: 节点图设计
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.SaveDesignRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/xresponse.Response'
      security:
      - BearerAuth: []
      summary: 保存流程设计
      tags:
      - 审批流程
  /approval/flows/{id}/enable:
    put:
      consumes:
      - application/json
      parameters:
      - description: 流程ID
        in: path
        name: id
        required: true
        type: integer
      - description: 启用状态
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/xresponse.Response'
      security:
      - BearerAuth: []
      summary: 启用/禁用流程
      tags:
      - 审批流程
  /approval/flows/by-type:
    get:
      parameters:
      - description: 表单类型(如 CONTRACT/EXPENSE/TRIP 等)
        in: query
        name: form_type
        required: true
        type: string
      - description: 表单标识(仅 OA_CUSTOM 按模板细分)
        in: query
        name: form_key
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/xresponse.Response'
      security:
      - BearerAuth: []
      summary: 按表单类型获取审批流程
      tags:
      - 审批流程
...
  /enterprise/job-logs:
    get:
      parameters:
      - description: 页码(默认1)
        in: query
        name: page
        type: integer
      - description: 每页条数(默认10,最大100)
        in: query
        name: page_size
        type: integer
      - description: 任务ID(可选过滤)
        in: query
        name: job_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/xresponse.Response'
      security:
      - BearerAuth: []
      summary: 任务执行日志
      tags:
      - 定时任务
  /enterprise/jobs:
    get:
      parameters:
      - description: 页码(默认1)
        in: query
        name: page
        type: integer
      - description: 每页条数(默认10,最大100)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/xresponse.Response'
      security:
      - BearerAuth: []
      summary: 定时任务列表
      tags:
      - 定时任务
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建任务请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.CreateSysJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/xresponse.Response'
      security:
      - BearerAuth: []
      summary: 创建定时任务
      tags:
      - 定时任务
  /enterprise/jobs/{id}:
    delete:
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/xresponse.Response'
      security:
      - BearerAuth: []
      summary: 删除定时任务
      tags:
      - 定时任务
    get:
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/xresponse.Response'
      security:
      - BearerAuth: []
      summary: 任务详情
      tags:
      - 定时任务
    put:
      consumes:
      - application/json
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新任务请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.UpdateSysJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/xresponse.Response'
      security:
      - BearerAuth: []
      summary: 更新定时任务
      tags:
      - 定时任务
  /enterprise/jobs/{id}/run:
    post:
      description: 立即执行一次指定任务(异步,不等待结果)
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/xresponse.Response'
      security:
      - BearerAuth: []
      summary: 手动触发任务
      tags:
      - 定时任务
...
  /system/homepage-config:
    get:
      description: 返回各板块开关 + 精选条目列表(含业务名称)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/xresponse.Response'
      security:
      - BearerAuth: []
      summary: 获取首页板块配置
      tags:
      - 站点设置
  /system/homepage-config/features/{id}:
    delete:
      parameters:
      - description: 精选条目ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/xresponse.Response'
      security:
      - BearerAuth: []
      summary: 删除单条精选条目
      tags:
      - 站点设置
  /system/homepage-config/sync:
    put:
      consumes:
      - application/json
      description: 全量替换某板块的精选条目(传入完整 ID 列表)
      parameters:
      - description: 精选条目
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.SyncFeaturesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/xresponse.Response'
      security:
      - BearerAuth: []
      summary: 同步精选条目
      tags:
      - 站点设置
  /system/homepage-config/toggle:
    put:
      consumes:
      - application/json
      description: 开启/关闭某板块在 CMS 首页的显示
      parameters:
      - description: 板块开关
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.ToggleModuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/xresponse.Response'
      security:
      - BearerAuth: []
      summary: 切换板块开关
      tags:
      - 站点设置
...
securityDefinitions:
  BearerAuth:
    description: 输入 "Bearer {access_token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"